==> Code/2_corpus.ipynb <==
# Corpus

___

Here we are going to clean up our datframe so that it is ready to feed into Count Vecotrizer and TFIDF vecotrizer.

import pickle
import pandas as pd

# Prep dataframe (df)

___

read df from reddit.pkl

with open('raw.pkl','rb') as raw_pickle:
    corpus = pickle.load(raw_pickle)
    

## Tokenizing

We will use this tokenizer to remove punctuation

from myfunct import *

Using the functions from myfunct.py we can prep our text by tokenizing, removing punctuation, and restriunging our text fields. 

corpus = make_clean_string_corpus(corpus,'selftext')

corpus.head()

It's important for us to keep this step seperate, so that we avoid making changes to our Raw.csv file. We don't want to have to go back and query the API again.

## Export

with open('corpus.pkl','wb') as clean_pickle:
    pickle.dump(corpus, clean_pickle)
