==> Code/5_pipe_forge.ipynb <==
# FORGE

___

Here we are going to do initial model testingon two selected models. K-Nearest Neighbors and Support Vector Machines.

We will fit and tune hyperparameters for those models in this notebook.

# Imports

___

import pickle
import pandas as pd
import numpy as np


from sklearn.pipeline import Pipeline 
from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, classification_report, mean_squared_error

from sklearn.model_selection import cross_val_score
from sklearn.naive_bayes import BernoulliNB, MultinomialNB
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.feature_extraction import text
from sklearn.svm import LinearSVC
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report

from gensim.models.word2vec import Word2Vec
import gensim



from nltk.tokenize import sent_tokenize, word_tokenize, RegexpTokenizer

import random
random.seed(42)

with open('corpus.pkl','rb') as clean_pickle:
    corpus = pickle.load(clean_pickle)

corpus.to_csv('./corpus.csv')

___

my_stop_words = ['https','com','www','people','know','actually',
                     'world','time','years','fact','facts','fake','like',
                     'sk','10','en','day','water','did','just','the']
    
    # append custom stopwords to text.ENGLIS_STOP_WORDS
stop_words = text.ENGLISH_STOP_WORDS.union(my_stop_words)
    

## Train Test Split

___

X = corpus['selftext']
y = corpus.subreddit

X_train, X_test, y_train, y_test = train_test_split(X,y,
                                                   stratify = y ,
                                                   test_size = .33,
                                                   random_state = 42)

# Logistic Regression 

___

## Cvec

logreg = Pipeline([
        ('cvec', CountVectorizer(stop_words=stop_words)),
        ('logreg',LogisticRegression(solver = 'saga'))]);



logreg_params = {
    'cvec__max_df': np.linspace(0.20,0.30,10),
    'cvec__min_df': np.linspace(0.001,1,10),
    'logreg__penalty': ['l1','l2'],
    'logreg__dual': [True,False],
    'logreg__C': np.linspace(0.001,1,10),
    #'logreg__solver': ['newton-cg','lbfgs','liblinear','sag','saga'],
    'logreg__max_iter': [5_000]
}    


grid = GridSearchCV(logreg, logreg_params, cv=5, n_jobs = 6, verbose = 2)

%time
grid.fit(X_train,y_train)

grid.best_estimator_

grid.best_score_

log_preds = grid.predict(X_test)

summary = classification_report(y_test,log_preds)

print(summary)

grid.score(y_test,grid.predict(X_test))

predictions_df = pd.DataFrame()

def _chart_grid(grid,name):
    temp_df = pd.DataFrame()
    probs = grid.predict_proba(X_test)
    preds = grid.predict(X_test)
    
    temp_df[f'{name} Probability'] = [prob[1] for prob in probs]
    temp_df[f'{name} Prediction'] = [pred for pred in log_preds]
    temp_df[f'{name} Binary'] = np.where(temp_df[f'{name} Prediction'] == 'facts',1,0)
    
    return temp_df

predictions_df =  _chart_grid(grid,'LogReg')

y_test = np.where(y_test == 'facts',1,0)

preds = predictions_df['LogReg Binary']

#preds = grid.predict(X_test)
mean_squared_error(y_test,preds)

predictions_df

## Support Vectors

{'svm__C': 0.112,
 'svm__max_iter': 1000,
 'tfidf__max_df': 0.223,
 'tfidf__min_df': 0.001,
 'tfidf__ngram_range': (1, 3),
 'tfidf__norm': 'l2'}

{'svm__C': 0.556,
 'svm__max_iter': 1000,
 'tfidf__max_df': 0.2,
 'tfidf__min_df': 0.0001,
 'tfidf__ngram_range': (1, 3),
 'tfidf__norm': 'l2'}



svm_tf = Pipeline([
        ('tfidf', TfidfVectorizer(stop_words=stop_words)),
        ('svm', LinearSVC() ) ]);

svm_tf_params = {
    'tfidf__max_df': np.linspace(0.20,0.30,10),
    'tfidf__min_df': np.linspace(0.0001,0.005,10),
    'tfidf__ngram_range': [(1,2),(1,3),(1,4)],
    'tfidf__norm' : ['l1','l2'],
    
    #'svm__loss': ['hinge','squared_hinge'],
    #'svm__dual': [True,False],
    'svm__C': np.linspace(0.001,1,10),
    'svm__max_iter': [1_000,2_000,3_000]
}    
    
grid_2 = GridSearchCV(svm_tf, svm_tf_params, cv=5, n_jobs = 6, verbose = 2)
    #grid = GridSearchCV(logreg, logreg_params, cv=5, n_jobs = 6, verbose = 2)
%time

grid_2.fit(X_train,y_train)

grid_2.best_score_

grid_2.best_params_

tokenizer = RegexpTokenizer('\w+|\$[\d\.]+|\S+')

logreg = Pipeline([
        ('tfidf', TfidfVectorizer(stop_words=stop_words)),
        ('MNB',MultinomialNB())]);



logreg_params = {
    'tfidf__max_df': np.linspace(0.20,0.30,10),
    'tfidf__min_df': np.linspace(0.001,1,10),
    'tfidf__tokenizer': [None,tokenizer],
    'tfidf__ngram_range':[(1,1),(1,2),(2,2)],
    'MNB__alpha': np.linspace(0.001,1,10)  
}    


grid_3 = GridSearchCV(logreg, logreg_params, cv=5, n_jobs = 6, verbose = 2)

%time
grid_3.fit(X_train,y_train)

grid_3.best_score_
